@@ -8,10 +8,12 @@ export class AuthGuard implements CanActivate {
   constructor(private auth: Auth, private router: Router) {}
 
   canActivate(route: ActivatedRouteSnapshot, state: RouterStateSnapshot) {
-      if (this.auth.authenticated()) {
+      if (this.auth.isAuthenticated()) {
+        console.log("it's true");
         return true;
       } else {
         this.router.navigate(['signup']);
+        console.log("it's false");
         return false;
       }
   }
